Distinguish between sgen and sfor, instead of scg everywhere
Now that I've done this, I'm actually wondering whether this is the good
approach. I think we might want to be more subtle than this. In
particular, the chain generation might want to keep talking about scg
because that is the (theoretical) parameter guiding the generation of
chains in real life. However, the implementation should definitely rely
on sgen and sfor and our classifiers should be in term of those two.
This commit is therefore more intended as food for thoughts rather than
anything else.